[IS-11] refactor and separate table component
