+++ src/components/TableComponents/CellTextItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CellTextItem";
