+++ src/components/TableComponents/HeadComponents/HeadComponents.tsx
@@ -1,0 +1,27 @@
+import { HeadRow, StyledHeadCell } from "./styled";
+
+interface IHeadComponentsProps {
+  columnsNames: [string, string, string, string];
+}
+
+const HeadComponents = ({ columnsNames }: IHeadComponentsProps) => {
+  return (
+    <HeadRow>
+      <StyledHeadCell $cellWidth="first" align="left">
+        {columnsNames[0]}
+      </StyledHeadCell>
+      <StyledHeadCell $cellWidth="second" align="left">
+        {columnsNames[1]}
+      </StyledHeadCell>
+      <StyledHeadCell $cellWidth="third" align="left">
+        {columnsNames[2]}
+      </StyledHeadCell>
+      <StyledHeadCell $cellWidth="fourth" align="left">
+        {columnsNames[3]}
+      </StyledHeadCell>
+      <StyledHeadCell $cellWidth="fifth" align="left"></StyledHeadCell>
+    </HeadRow>
+  );
+};
+
+export default HeadComponents;
